+++ kernel-rs/src/arch/mod.rs
@@ -1,5 +1,7 @@
 //! Architecture-dependent code.
 
+pub mod addr;
 pub mod memlayout;
 pub mod plic;
+pub mod poweroff;
 pub mod riscv;
