+++ kernel-rs/src/arch/poweroff.rs
