+++ kernel-rs/src/console.rs
@@ -1,12 +1,12 @@
 use core::fmt;
 
 use crate::{
+    arch::addr::UVAddr,
     file::Devsw,
     kernel::{kernel, kernel_builder},
     lock::SleepablelockGuard,
     param::NDEV,
     uart::Uart,
-    vm::UVAddr,
 };
 
 const CONSOLE_IN_DEVSW: usize = 1;
