+++ kernel-rs/src/fs/inode.rs
@@ -78,6 +78,7 @@ use static_assertions::const_assert;
 
 use super::{FileName, Stat, IPB, MAXFILE, NDIRECT, NINDIRECT};
 use crate::{
+    arch::addr::UVAddr,
     arena::{Arena, ArenaObject, ArrayArena, Rc},
     bio::BufData,
     fs::{FsTransaction, Path, ROOTINO},
@@ -86,7 +87,6 @@ use crate::{
     param::ROOTDEV,
     param::{BSIZE, NINODE},
     proc::CurrentProc,
-    vm::UVAddr,
 };
 
 /// Directory is a file containing a sequence of Dirent structures.
