+++ kernel-rs/src/lib.rs
@@ -70,12 +70,11 @@ mod lock;
 mod page;
 mod param;
 mod pipe;
-mod poweroff;
 mod proc;
 mod start;
 mod syscall;
 mod trap;
 mod uart;
-mod utils;
+mod util;
 mod virtio;
 mod vm;
