+++ kernel-rs/src/syscall/mod.rs
@@ -3,10 +3,10 @@ use core::{mem, str};
 use cstr_core::CStr;
 
 use crate::{
+    arch::addr::{Addr, UVAddr},
     kernel::Kernel,
     println,
     proc::CurrentProc,
-    vm::{Addr, UVAddr},
 };
 
 mod file;
