+++ kernel-rs/src/util/etrace.rs
