+++ kernel-rs/src/util/list.rs
