+++ kernel-rs/src/util/mod.rs
