+++ kernel-rs/src/util/rc_cell.rs
