+++ kernel-rs/src/virtio/virtio_disk.rs
@@ -15,7 +15,7 @@ use super::{
     VIRTIO_BLK_T_IN, VIRTIO_BLK_T_OUT,
 };
 use crate::{
-    arch::riscv::{PGSHIFT, PGSIZE},
+    arch::addr::{PGSHIFT, PGSIZE},
     bio::Buf,
     kernel::kernel_builder,
     lock::{Sleepablelock, SleepablelockGuard},
